pyproject.toml
[tool.poetry]
name = "babbler"
version = "0.1.0"
description = "A CLI app for interacting with generative APIs"
authors = ["Gerard <gerardnoth@gmail.com>"]
maintainers = ["Gerard <gerardnoth@gmail.com>"]
readme = "README.md"
packages = [{ include = "babbler", from = "src" }]
license = "MIT"
repository = "https://github.com/gerardnoth/babbler"

[tool.poetry.dependencies]
python = "~3.12"
google-cloud-aiplatform = "1.65.0"
google-generativeai = "0.7.2"
loguru = "0.7.2"
openai = "1.44.1"
orjson = "3.10.7"
pydantic = "2.8.2"
python-dotenv = "1.0.1"
tqdm = "4.66.5"
typer = "0.12.5"

[tool.poetry.group.dev.dependencies]
mypy = "1.11.2"
pytest = "8.3.2"
pytest-httpserver = "1.1.0"
ruff = "0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = [
    "--doctest-modules",
    "--import-mode=importlib",
]

[tool.mypy]
# mypy global options: https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true
